+++ aidial_adapter_bedrock/dial_api/resource.py
